refactor: move terrain store into core service
